pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "lxml>=6.0.0",
    "python-telegram-bot>=22.2",
    "requests>=2.32.4",
    "trafilatura>=1.6.1",
]
